ARM: tegra: clock: fix build warnings
Remove warning 
-initialization from incompatible pointer type

Bug 1211919

Reviewed-on: http://git-master/r/411952
Tested-by: Sachin Nikam <snikam@nvidia.com>

+++ arch/arm/mach-tegra/clock.c
@@ -1453,7 +1453,7 @@ static void syncevent_one(struct clk *c)
 	}
 }
 
-static int syncevent_write(struct file *file, const char __user *user_buf,
+static ssize_t syncevent_write(struct file *file, const char __user *user_buf,
 				size_t count, loff_t *ppos)
 {
 	struct clk *c;
